Bug #1719430 devfreq: sprd: fix ddr_dvfs driver pc-lint warning
[root cause  ] pc-lint warning
[changes     ] fix ddr_dvfs driver pc-lint warning
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:sharkl5pro
[test case]:read/write sysfs
[reviewers   ]
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

+++ drivers/devfreq/sprd/sprd_ddr_dvfs_sp.c
@@ -175,7 +175,8 @@ static int dvfs_probe(struct platform_device *pdev)
 		dev_warn(dev, "failed read scene_num\n");
 		scene_num = 0;
 	} else {
-		g_vote_data = devm_kzalloc(dev, sizeof(struct vote_data)*scene_num, GFP_KERNEL);
+		g_vote_data = devm_kzalloc(dev, sizeof(struct vote_data)*(unsigned int)scene_num,
+					   GFP_KERNEL);
 		if (g_vote_data == NULL) {
 			err = -ENOMEM;
 			return err;
